% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_dist}
\alias{calc_dist}
\alias{calc_dist.likefree_model}
\alias{calc_dist.sim_results}
\title{Calculate the distance between data and simulated dynamics}
\usage{
calc_dist(sim, pars, ...)

\method{calc_dist}{likefree_model}(sim, pars)

\method{calc_dist}{sim_results}(sim)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Calculate the distance between data and simulated dynamics
}
