% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_likefree_model}
\alias{compile_likefree_model}
\title{Compiling a likelihood-free simulation model}
\usage{
compile_likefree_model(
  dat,
  m_sim,
  y0,
  rprior,
  dprior,
  times = dat[, "t"],
  m_warmup = NA,
  t_warmup = 100
)
}
\arguments{
\item{dat}{a dataframe of data to be fitted, "t" as the indicator of time}

\item{m_sim}{an odin model}

\item{y0}{initial values of the model}

\item{rprior}{a function for simulating prior parameters}

\item{dprior}{a function for calulating prior probability in log scale}

\item{times}{a vectors of times for projecting resutls}

\item{m_warmup}{a odin model for warm-up period}

\item{t_warmup}{the length of warm-up period}
}
\value{

}
\description{
Compiling a likelihood-free simulation model
}
