% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model_likelihood}
\alias{compile_model_likelihood}
\title{Compile a simulation model with a likelihood function to data}
\usage{
compile_model_likelihood(fn_like, sim)
}
\arguments{
\item{fn_like}{likelihood function}

\item{sim}{a compile model, see compile_model}
}
\value{
a compiled model with likelihood function
}
\description{
Compile a simulation model with a likelihood function to data
}
