% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{ess}
\alias{ess}
\alias{ess.numeric}
\alias{ess.fitted_abc}
\title{Effective Sample Size (ESS)}
\usage{
ess(obj)

\method{ess}{numeric}(obj)

\method{ess}{fitted_abc}(obj)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Effective Sample Size (ESS)
}
