% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit simulation model to data with Likelihood-free methods}
\usage{
fit(lf, n_posterior, method = c("abc", "abcsmc", "abcpmc"), ...)
}
\arguments{
\item{...}{arguments for a specific method}
}
\value{

}
\description{
Fit simulation model to data with Likelihood-free methods
}
