% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{fit_abc}
\alias{fit_abc}
\alias{summary.fitted_abc}
\title{Approximate Bayesian Computation (ABC) for System dynamic models}
\usage{
fit_abc(
  sim,
  n_posterior = 300,
  epsilon = NA,
  keep_all = FALSE,
  target_acc = 0.05
)

\method{summary}{fitted_abc}(fitted)
}
\arguments{
\item{target_acc}{}
}
\value{

}
\description{
Approximate Bayesian Computation (ABC) for System dynamic models
}
