% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcpmc.R
\name{fit_abc_pmc}
\alias{fit_abc_pmc}
\title{Title}
\usage{
fit_abc_pmc(
  lf,
  n_posterior,
  k = 5,
  keep = c("Y0", "Ys", "both", "none"),
  max_round = 20,
  q_stop = 0.99,
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
\references{
Simola, U., Cisewski-Kehe, J., Gutmann, M. U., & Corander, J. (2019). Adaptive Approximate Bayesian Computation Tolerance Selection. arXiv preprint arXiv:1907.01505.
}
