% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lse}
\alias{lse}
\title{LogSumExp function}
\usage{
lse(xs, na.rm = FALSE)
}
\arguments{
\item{xs}{a vector of values to be summed up in exp scale}
}
\value{
a value
}
\description{
LogSumExp function
}
\examples{
xs <- c(1000, 1001, 999)

## go infinity because of numerical overflow
is.infinite(log(sum(exp(x))))

## Use lse function
lse(xs)
}
