% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\alias{project.likefree_model}
\title{Projecting from a simulation model}
\usage{
project(sim, y_eq, pars, ...)

\method{project}{likefree_model}(sim, y_eq, pars, times = NA)
}
\arguments{
\item{...}{}
}
\value{
a data.frame of simulated data
}
\description{
Projecting from a simulation model
}
