% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.likefree_model}
\alias{simulate.likefree_model}
\title{Simulate a system dynamic}
\usage{
\method{simulate}{likefree_model}(sim, y0 = NA, pars = NA, times = NA)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Simulate a system dynamic
}
