% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmup.R
\name{warmup}
\alias{warmup}
\alias{warmup.sim_model}
\alias{update.Y_eq}
\title{Warming-up a model to a steady state}
\usage{
warmup(sim, ...)

\method{warmup}{sim_model}(sim, y0, pars, times = sim$TS_wp)

\method{update}{Y_eq}(yeq, sim, nforward = length(sim$TS_wp))
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Warming-up a model to a steady state
}
