% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{test_model}
\alias{test_model}
\alias{compile_model}
\title{Compile a simulation model given all components}
\usage{
test_model(
  d_prior,
  r_prior,
  y0,
  inp_sim,
  ts_sim,
  m_sim,
  inp_wp,
  t_wp,
  m_wp,
  fn_pass_y0,
  fn_check
)

compile_model(
  d_prior,
  r_prior,
  y0,
  inp_sim = NULL,
  ts_sim,
  m_sim,
  inp_wp = NULL,
  t_wp,
  m_wp = m_sim,
  fn_pass_y0,
  fn_check,
  max_attempt = 10
)
}
\arguments{
\item{d_prior}{a probability density function supporting prior distributions}

\item{r_prior}{a function for generating parameters from their prior distributions}

\item{y0}{initial values}

\item{inp_sim}{input data for the simulation stage}

\item{ts_sim}{timespan for simulation}

\item{m_sim}{an odin model for simulation}

\item{inp_wp}{input data for the warm-up stage}

\item{t_wp}{length of warm-up stage}

\item{m_wp}{an odin model for waru-up}

\item{fn_pass_y0}{a function for bringing the states at the end of warm-up to simulation initials}

\item{fn_check}{a function for checking if a parameter set can generate validated output}
}
\value{

}
\description{
Compile a simulation model given all components
}
