% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model_likefree}
\alias{compile_model_likefree}
\title{Compile a simulation model with a likelihood-free link to data}
\usage{
compile_model_likefree(dat, sim)
}
\arguments{
\item{dat}{a dataframe of data to be fitted, "t" as the indicator of time}

\item{sim}{a compile model, see compile_model}
}
\value{
a compiled model with data
}
\description{
Compile a simulation model with a likelihood-free link to data
}
