% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit simulation model to data with Likelihood-free methods}
\usage{
fit(lf, n_posterior, method = c("abc", "abcsmc", "abcpmc"), ...)
}
\arguments{
\item{lf}{a sim_model_likefree object}

\item{n_posterior}{number of posterior particles to collect}

\item{method}{algorithm for model fitting}

\item{...}{arguments for a specific method}
}
\value{

}
\description{
Fit simulation model to data with Likelihood-free methods
}
