% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{fit_abc}
\alias{fit_abc}
\title{Approximate Bayesian Computation (ABC) for System dynamic models}
\usage{
fit_abc(lf, n_posterior = 300, epsilon = NA, target_acc = 0.1, n_test = 100)
}
\arguments{
\item{lf}{a "model_likefree" object with a compile model and data}

\item{n_posterior}{number of posteriors to collect}

\item{epsilon}{threshold for collecting}

\item{target_acc}{target acceptance rate}

\item{n_test}{number of simulations for testing threshold}
}
\value{

}
\description{
Approximate Bayesian Computation (ABC) for System dynamic models
}
\examples{
library(odin2data)

## Example 1: model without warmup stage
test_data <- data.frame(
  t = 1:5,
  incidence = c(20, 49, 109, 184, 206) / 1000
)

### Load the model file
f <- system.file("models/SIR.txt", package = "odin2data")
test_m <- odin::odin(f, verbose=F)


### Generate a prior and set up it as a list
r_prior <- function() {
  list(beta = runif(1, 1, 10), gamma = runif(1, .1, 1))
}

d_prior <- function(pars) {
  dunif(pars$beta, 1, 10, log = T) + dunif(pars$gamma, .1, 1, log = T)
}

times = seq(0, 10, 0.2)
y0 <- c(995, 5, 0)

### Compile all elements as a simulation model
sim <- odin2data::compile_model(d_prior, r_prior, y0, ts_sim = times, m_sim = test_m)

### Compile the model with data
lf <- odin2data::compile_model_likefree(test_data, sim)

### Model fitting
post <- odin2data::fit_abc(lf, 100, target_acc = 0.2)
summary(post)


## Example 2: model with warmup stage
fn_pass_y0 <- function(ys) {
ys[nrow(ys), 2:4]
}

test_data <- data.frame(
  t = 1:5,
  incidence = rep(0.009580023, 5)
)

sim <- odin2data::compile_model(d_prior, r_prior, y0, ts_sim = times, m_sim = test_m,
                                m_wp = test_m, t_wp = 100, fn_pass_y0 = fn_pass_y0)

lf <- odin2data::compile_model_likefree(test_data, sim)

### Model fitting
post <- odin2data::fit_abc(lf, 100, target_acc = 0.2)
summary(post)

}
