% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcsmc.R
\name{fit_abc_smc}
\alias{fit_abc_smc}
\title{Approximate Bayesian Computation Sequential Monte Carlo}
\usage{
fit_abc_smc(lf, n_posterior, alpha = 0.9, p_thres = 0.6,
  max_round = 20, max_stay = 3, verbose = T)
}
\arguments{
\item{lf}{a sim_model_likefree object}

\item{n_posterior}{number of posterior particles to collect}

\item{alpha}{percentage particles surviving}

\item{p_thres}{proportion of threshold population}

\item{max_round}{max round, reach to stop}

\item{max_stay}{max staying round of epsilon, reach to stop}

\item{verbose}{show log or not}

\item{keep}{keep simulation (Ys), intial values (Y0), or both}
}
\description{
Approximate Bayesian Computation Sequential Monte Carlo
}
\references{
Del Moral, P., Doucet, A., & Jasra, A. (2012). An adaptive sequential Monte Carlo method for approximate Bayesian computation. Statistics and Computing, 22(5), 1009-1020.
}
