% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.sim_model}
\alias{simulate.sim_model}
\alias{simulate.sim_model_likefree}
\title{Simulate a single run}
\usage{
\method{simulate}{sim_model}(sim, y0, pars, warmup = F)

\method{simulate}{sim_model_likefree}(sim, y0, pars, warmup = F)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Simulate a single run
}
