% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmup.R
\name{warmup}
\alias{warmup}
\alias{warmup.likefree_model}
\title{Warming-up a model to a steady state}
\usage{
warmup(sim, ...)

\method{warmup}{likefree_model}(sim, y0 = NA, pars, t_warmup = NA)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Warming-up a model to a steady state
}
